Main and Global classes were exracted in their seperate java files.
