+++ src/GeneticAlgorithm.java
@@ -1,3 +1,4 @@
+package src;
 
 /*========================================================================
  =                                                                       =
