+++ src/GeneticAlgorithmTrainer.java
@@ -1,3 +1,4 @@
+package src;
 /*========================================================================
  =                                                                       =
  =     The Game Fifteen 0.51                                             =
