+++ src/Globals.java
@@ -1,0 +1,16 @@
+package src;
+
+/**
+ * Global constants class.
+ * 
+ * @author Todor Balabanov
+ * @author tdb@tbsoft.eu
+ */
+class Globals {
+	public static final int RIGHT = 1;
+	public static final int DOWN = 2;
+	public static final int LEFT = 4;
+	public static final int UP = 16;
+
+	public static final int EMPTY = 16;
+}
